Create base classes for WSGI apps
We created parent classes `BaseWSGIApp` and `ObjectStoreWSGIApp`
for discovery/registry/repository app classes.
Now we can reuse methods defined in parent classes
and avoid code duplication.
